refactor: extend mxCellRender instead of redefining prototypes
This makes things more explicit and avoid side effects.
